==> notebooks/data-exploration/01_modify_data.ipynb <==
import numpy as np
from matplotlib import pyplot as plt
import seaborn as sns
import pandas as pd
import json
import re

# #############
#   load data
# #############
path_to_open = "..\\data\\_properties_data.json"

data = None
with open(path_to_open, "r", encoding="utf-8") as file:
    data = json.load(file)

df = pd.DataFrame.from_dict(data, orient='index')

df.head()

# ######################
#   fix locality names
# ######################

# load new locality data
ref_data = pd.read_csv("..\data\_postal_code_data.csv")
ref_data.head()

zip = 9050
ref = ref_data.query("postalCode == @zip")
display(ref)

"""
for index, row in df.iterrows():
    pcode = int(row["postalCode"])
    locality = row["locality"]
    pcode_data = ref_data.query('postalCode == @pcode')
    for ix, rw in pcode_data.iterrows():
        loc = locality.lower()
        cit = rw["city"].lower()
        mun = rw["municipality"].lower()
        can = None
        if type(rw["cityAltName"]) != float:
            can = rw["cityAltName"].lower()
        name_found = False
        if loc.find(cit) != -1 or loc.find(mun) != -1 or mun.find(loc) != -1:
            name_found = True
        if name_found:
            df.loc[index, "city"] = rw["city"]
            df.loc[index, "municipality"] = rw["municipality"]
            continue
        else:
            df.loc[index, "city"] = rw["city"]
            df.loc[index, "municipality"] = rw["municipality"]
"""

#df[["postalCode", "locality", "latitude", "longitude", "city", "municipality"]][df["postalCode"] == "9050"].head(10)
df[["postalCode", "locality", "latitude", "longitude"]][df["postalCode"] == "9050"].head(10)

df.isna().sum()

# #####################
#   add locality type
# #####################
def addLocalityType(postalCode):
    localityType = 0
    # X000 & X00X > 1
    if re.match(r"^[1-9][0]{3}$", postalCode) or re.match(r"^[1-9][0]{2}[1-9]$", postalCode):
        localityType = 1
    # X0X0 & X0XX> 2
    elif re.match(r"^([1-9][0]){2}$", postalCode) or re.match(r"^[1-9][0][1-9]{2}$", postalCode):
        localityType = 2
    # XX00 & XX0X > 3
    elif re.match(r"^[1-9]{2}[0]{2}$", postalCode) or re.match(r"^[1-9]{2}[0][1-9]$", postalCode):
        localityType = 3
    # XXX0 & XXXX > 4
    elif re.match(r"^[1-9]{3}[0]$", postalCode) or re.match(r"^[1-9]{4}$", postalCode):
        localityType = 4
    return localityType

df["localityType"] = df["postalCode"].apply(addLocalityType)
df[["postalCode", "localityType"]].head()


df.groupby("localityType")["type"].count()

# #################
#   merge address
# #################
df["address"] = df["street"] + " " + df["number"] #+ ", " + df["box"]
df[["street", "number", "box", "address"]].head()

# ###################
#   price per meter  
# ###################
df["pricePerMeter"] = round(df["price"] / df["netHabitableSurface"])
df[["price", "netHabitableSurface", "pricePerMeter"]].head()

# ################################################
#   apply new order and remove redundant columns
# ################################################
columns_to_take = [
    "type", "subtype", 
    "region", "province", "district", "locality", "localityType", 
    "postalCode", "address", "latitude", "longitude", 
    "condition", "constructionYear", "price", "pricePerMeter", "bedroomCount", "netHabitableSurface", 
    "floor", "floorCount", "hasLift", "facadeCount", 
    "hasGarden", "gardenSurface", "hasTerrace", "terraceSurface", "land", 
    "kitchen", "bathroomCount", "showerRoomCount", "toiletCount", 
    "fireplaceExists", "hasSwimmingPool", "hasAirConditioning", 
    "parkingCountIndoor", "parkingCountOutdoor", "parkingCountClosedBox", 
    "primaryEnergyConsumptionPerSqm", "epcScore", "hasDoubleGlazing", 
    "saleType" 
    ]
df_reduced = df[columns_to_take]
df_reduced.head()


# ################
#   save new csv
# ################
path_to_save = "..\\data\\01_modifyData_properties_data.csv"
df_reduced.to_csv(path_to_save, index=True, index_label="ID")
