==> notebooks/data-exploration/02_clean_data.ipynb <==
import numpy as np
from matplotlib import pyplot as plt
import seaborn as sns
import pandas as pd
import json

# #############
#   load data
# #############
df = pd.read_csv("..\\data\\01_modifyData_properties_data.csv", index_col="ID")
df.shape

# ###########################
#   remove "bad" properties
# ###########################


# type of sale
saleType2drop = ['NewRealEstateProject', 'NotarySale', 'PublicSale', 'LifeAnnuitySale', 'InvestmentProject']
for saleType in saleType2drop:
    df.drop(df[df['saleType'] == saleType].index, inplace=True)
print(df.shape)

# remove some bad values
# bedrooms
df.drop(df[df["bedroomCount"] < 1].index, inplace=True)
df.drop(df[df["bedroomCount"] > 30].index, inplace=True)


# livig surface
df.drop(df[df["netHabitableSurface"] < 20].index, inplace=True)
df.drop(df[df["netHabitableSurface"] > 1000].index, inplace=True)
#df.groupby("netHabitableSurface")["type"].count()

df.groupby("netHabitableSurface")["type"].count()

# price
df.drop(df[df["price"] < 10000].index, inplace=True)


epc = ['G_A++', 'C_B', 'G_F', 'G_A', 'E_A', 'D_C']
df["epcScore"] = df["epcScore"].str.replace(r"^[A-Z]_", "", regex=True)
df["epcScore"] = df["epcScore"].str.replace("+", "")
df.groupby("epcScore")["type"].count()


# print shape
print(df.shape)

# ############################
#   remove redundant columns
# ############################


# #####################
#   remove duplicates
# #####################


# ##############################
#   reduce parking information
# ##############################


df.info()

# ################
#   save new csv
# ################
path_to_save = "..\\data\\02_cleanData_properties_data.csv"
df.to_csv(path_to_save, index=True)
